Consolidate the two Plot extension files we had.
Also, Components is now a loaded word in Plot vernacular.
